+++ CHANGELOG.rst
@@ -1,7 +1,7 @@
 1.5.5
 -----
 
-
+* Fixed :code:`'NoneType' object has no attribute 'groupdict'` in :meth:`.TimeDelta.cast`. (#656)
 
 1.5.4 - December 27, 2016
 -------------------------